mds/mdstypes: move old_inode_t to separate header
Reduce header dependencies of CInode.h.
