+++ src/mds/CInode.h
@@ -37,7 +37,7 @@
 #include "include/object.h" // for object_t
 
 #include "MDSCacheObject.h"
-#include "mdstypes.h" // for old_inode_t
+#include "old_inode.h" // for old_inode_t
 #include "flock.h"
 #include "inode_backtrace.h" // for inode_backtrace_t
